+++ src/main/java/de/samply/security/SecurityConfiguration.java
@@ -83,6 +83,7 @@ public class SecurityConfiguration {
 
         // Services without authentication required
         authorization.requestMatchers(HttpMethod.GET, ProjectManagerConst.INFO).permitAll();
+        authorization.requestMatchers(HttpMethod.GET, ProjectManagerConst.FETCH_FRONTEND_VARIABLES).permitAll();
         authorization.requestMatchers(HttpMethod.GET, ProjectManagerConst.ASSETS).permitAll();
 
         // Services with authentication required