+++ CHANGELOG.md
@@ -270,7 +270,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Condition in form field config
 - External assets for frontend
 - Form field as file
-- Form Field upload and download
+- Form field upload and download
+- Frontend variables
 
 
 ### Changed